BaseExecutor：把Executor改为BaseExecutor，"*.xxx"格式执行器将继承该父类
